Added file to manage random functions.  Needed to be separated from other code.
Makefile output generates executable named StockScreen instead of stock.

DrawString is back and improved.

Function IconArrow draws a red or green arrow to indicate direction of
change for data.

Removed many comments about stock market functionality.

Created DrawMovingList allows user to create a list of data that is
listed in numerical order.  List displays the movement of data points
over time.

DrawOther is a temporary function to encapsulate OpenGL code that is
not needed for this version of the program. This code will be used later.

DrawMovingList has a struct associated with it.  One has been created as
an example with some data for the user to view.

Main window has been set to a different size.